Added WeeklyCalendar, changed reprs
